refactor Badges display and Update ProfileImage Modal
